backend/requirements.txt
aio-pika==9.4.0
aiormq==6.8.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
multidict==6.0.5
pamqp==3.3.0
pika==1.3.2
pydantic==2.6.4
pydantic_core==2.16.3
pydantic-settings==2.2.1
python-telegram-bot==20.8
python-multipart==0.0.9
pyjwt==2.8.0
gspread-asyncio==2.0.0
sniffio==1.3.0
SQLAlchemy==1.4.52
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
yarl==1.9.4
beanie==1.25.0